requirements.txt
# GUI框架
dearpygui>=1.9.0

# HTTP客户端
httpx>=0.25.0

# 配置管理
python-dotenv>=1.0.0
pyyaml>=6.0

# 数据模型
pydantic>=2.0.0

# 测试
pytest>=7.0.0
pytest-asyncio>=0.21.0
